[Ticker] Adds all the ticker endpoints
I added a structure and did some abstractions that I believe to be
common place in a REST API. If they prove to be wrong, we can roll them
back.

I added the concept of a multi find, because the finding a ticker by id
returns an array, which is not the norm, and this is not the regular
find.