Completely changes the structure of the lib
Due to my misunderstanding of the API, I was going with a solution which
realied greatly in class methods. Turns out that we're going to need the
user's token for each request, and it is probably going to be different
at each request.

This means that we need a Bitreserve::Client class that we instantiate
with that token and go on from there.

My approach was to make the client have a huge public API with all of
the methods available (provided their are being mixed in).

I also moved those mixins into a directory called api and moved the
actual classes into a directory called entities.