[Cards] Support all endpoints for cards
Extract all of the request data into a value object. This makes it
easier to extend the data we pass to the request, and limits the number
of arguments that it needs on initialization.

Moves the authorization header method to the client, closer to where the
token info is. This removes the need to pass the client around.

Extract test logic for both Request methods into a shared example.

Closes #5.